requirements.txt
python-telegram-bot==21.0
pillow==10.2.0
